counter: stm32-timer-cnt: Fix stm32_count_direction_read return type
The counter_count_ext read callback requires a function with a ssize_t
return type. The stm32_count_direction_read function should return
ssize_t rather than int in order to match the counter_count_ext read
callback requirement.

+++ drivers/counter/stm32-timer-cnt.c
@@ -127,7 +127,7 @@ static int stm32_count_function_set(struct counter_device *counter,
 	return 0;
 }
 
-static int stm32_count_direction_read(struct counter_device *counter,
+static ssize_t stm32_count_direction_read(struct counter_device *counter,
 				      struct counter_count *count,
 				      void *private, char *buf)
 {
